pyproject.toml
[project]
name = "specter"
version = "0.1.0"
description = "SPECTER - AI-Powered Offensive Security Terminal"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SPECTER Team"}
]
keywords = ["security", "pentesting", "ai", "llm", "cybersecurity"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Security",
]

dependencies = [
    "rich>=13.7.0",
    "typer>=0.9.0",
    "pydantic>=2.5.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "prompt_toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
ollama = [
    "ollama>=0.1.0",
]
export = [
    "weasyprint>=60.0",
    "python-docx>=1.0.0",
]
workflows = [
    "pyyaml>=6.0",
]

[project.scripts]
specter = "specter.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
